+++ CHANGES.md
@@ -6,6 +6,7 @@ History
 * Add `pyppeteer-install` command
 * Add `autoClose` option to `launch` fucntion
 * Add `loop` option to `launch` function (experimental)
+* Add `Page.setBypassCSP` method
 * Continue to find WS URL while process is alive
 
 ## Version 0.0.20 (2018-08-11)